pyproject.toml
[project]
name = "xmlrolevalidator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fitz>=0.0.1.dev2",
    "fuzzywuzzy>=0.18.0",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.0",
    "pinecone-client>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
]
